refactor hles extraction
