+++ dap-etl/extraction/src/main/scala/org/broadinstitute/monster/dap/hles/HLESurveyExtractionPipeline.scala
@@ -1,17 +1,9 @@
-package org.broadinstitute.monster.dap
+package org.broadinstitute.monster.dap.hles
 
-import org.broadinstitute.monster.dap.common.{
-  Args,
-  ExtractionPipelineBuilder,
-  FilterDirective,
-  FilterOps,
-  HttpWrapper,
-  OkWrapper,
-  RedCapClient
-}
+import org.broadinstitute.monster.common.{PipelineBuilder, ScioApp}
+import org.broadinstitute.monster.dap.common._
 
 import java.time.{OffsetDateTime, ZoneOffset}
-import org.broadinstitute.monster.common.{PipelineBuilder, ScioApp}
 
 // Ignore IntelliJ, this is used to make the implicit parser compile.
 import Args._
